+++ Assets/ImportantPrototype/Scripts/Mutations/GenotypeMods/MultiplyNegativeEffectGenotypeMod.cs
@@ -7,7 +7,7 @@ namespace ImportantPrototype.Mutations
     public class MultiplyNegativeEffectGenotypeMod : GenotypeMod
     {
         [SerializeField]
-        private int _multiplier; 
+        private float _multiplier; 
         
         public override string GetDescription()
         {
