+++ Assets/ImportantPrototype/Scripts/Mutations/GenotypeMods/MultiplyPositiveEffectGenotypeMod.cs
@@ -7,7 +7,7 @@ namespace ImportantPrototype.Mutations
     public class MultiplyPositiveEffectGenotypeMod : GenotypeMod
     {
         [SerializeField]
-        private int _multiplier; 
+        private float _multiplier; 
         
         public override string GetDescription()
         {
